requirements.txt
ffmpeg-python>=0.2.0
Pillow>=9.0.0
pytesseract>=0.3.10
dataclasses>=0.6; python_version < "3.7" 